main page 재작업 복구
이전 작업에 실수 한 부분이 있어서 코드 다시 최신화 하였습니다
