+++ lib/views/Main_page.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:suwon/views/widgets/cancel_and_action_alert_dialog.dart';
-import 'package:suwon/views/edit_profie.dart';
+import 'package:suwon/views/EditProfileDrawerScreen.dart';
+import 'package:suwon/views/TermsConditionsDrawerScreen.dart';
+import 'package:suwon/views/FeedbackDrawerScreen.dart';
 import 'package:suwon/views/login_screen.dart';
 import 'package:suwon/views/matching_loding.dart';
 
@@ -44,14 +46,178 @@ class MainPage extends StatelessWidget {
                   mainAxisAlignment: MainAxisAlignment.end,
                   children: [
                     Container(
-                      child: SvgPicture.asset('asstes/onclikmenu.svg', width: 28.w,
-                          height: 20.h),
+                      child: IconButton(
+                        onPressed: (){
+                        _scaffoldKey.currentState?.openEndDrawer();
+                      },
+                          icon: SvgPicture.asset('asstes/onclikmenu.svg', width: 28.w,
+                          height: 20.h),)
                     ),
-                    SizedBox(width: 32.w,),
+                    SizedBox(width: 25.w,),
                   ],
                 ),
               ),
             ],
+          ),
+          endDrawer:
+          Container(
+            child:
+            Drawer(
+                backgroundColor: Colors.white,
+                width: 240.w,
+                child: Container(
+                  width: 240.w,
+                  child:
+                  ListView(
+                    physics: const NeverScrollableScrollPhysics(),
+                    children: [
+                      Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          DrawerHeader(
+                              margin: EdgeInsets.zero,
+                              child: Container(
+                                height: 291.h,
+                                child: Column(
+                                  children: [
+
+                                    Row(
+                                      mainAxisAlignment: MainAxisAlignment.end,
+                                      children: [
+                                        IconButton(onPressed: (){}, icon: Image.asset('asstes/close.png', height: 16.h, width: 16.h,))
+                                      ],
+                                    )
+                                  ],
+                                ),
+                              )
+                          ),
+
+                          Container(
+                            height: 402.h,
+                            width: double.infinity,
+                            decoration: BoxDecoration(
+                              border: Border(
+                                top: BorderSide(
+                                    color: Color(0xffEDEDED),
+                                    width: 1.w),
+                                bottom: BorderSide(
+                                  color: Color(0xffEDEDED),
+                                  width: 1.w,
+                                ),
+                              ),
+                            ),
+                            child: Column(
+                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                              children: [
+                                Container(
+                                  height: 397.h,
+                                  child: Column(
+                                    mainAxisAlignment: MainAxisAlignment.center,
+                                    children: [
+                                      ListTile_main(
+                                        text: '프로필 수정',
+                                        Menu_Icon: 'asstes/profile.svg',
+                                        onTap: ()
+                                        {
+                                          Navigator.push(
+                                            context,
+                                            MaterialPageRoute(
+                                              builder: (context) =>
+                                                  EditProfileDrawerScreen(),
+                                            ),
+                                          );
+                                        }, ),
+                                      ListTile_main(
+                                        text: '이용약관',
+                                        Menu_Icon: 'asstes/policy.svg',
+                                        onTap: () {
+                                          Navigator.push(
+                                            context,
+                                            MaterialPageRoute(
+                                              builder: (context) =>
+                                                  TermsConditionsDrawerScreen(),
+                                            ),
+                                          );
+                                        },
+                                      ),
+                                      ListTile_main(
+                                        text: '피드백',
+                                        Menu_Icon: 'asstes/feedback.svg',
+                                        onTap: () {
+                                          Navigator.push(
+                                            context,
+                                            MaterialPageRoute(
+                                              builder: (context) =>
+                                                  FeedbackDrawerScreen(),
+                                            ),
+                                          );
+                                        },
+                                      ),
+                                      ListTile_main(
+                                          text: '로그아웃',
+                                          Menu_Icon: 'asstes/logout.svg', onTap: () {
+                                        Navigator.push(
+                                          context,
+                                          MaterialPageRoute(
+                                              builder: (context) => NewLoignMain()),
+                                        );
+                                      }
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                          Container(
+                            height: 150.h,
+                            width: double.infinity,
+                            color: Color(0xffEDEDED),
+                            child: Row(
+                              children: [
+                                SizedBox(
+                                  width: 32.w,
+                                ),
+                                Column(
+                                  children: [
+                                    Column(
+                                      crossAxisAlignment: CrossAxisAlignment.start,
+                                      children: [
+                                        SizedBox(
+                                          height: 15.h,
+                                        ),
+                                        Image.asset(
+                                          'asstes/SUCHAT.png',
+                                          scale: 3,
+                                        ),
+                                        SizedBox(
+                                          height: 12.h,
+                                        ),
+                                        Text(
+                                          'Copyright 2023.\nFlag inc. all rights reserved.',
+                                          style: TextStyle(
+                                            color: Color(0xff767676),
+                                            fontFamily: 'Pretendard',
+                                            fontSize: 10.sp,
+                                            fontStyle: FontStyle.normal,
+                                            fontWeight: FontWeight.w400,
+                                            letterSpacing: 0.25.sp,
+                                          ),
+                                        ),
+                                        resign_Button(),
+                                      ],
+                                    )
+                                  ],
+                                )
+                              ],
+                            ),
+                          )
+                        ],
+                      )
+                    ],
+                  ),
+                )
+            ),
           ),
           body: Column(
             children: [
@@ -178,6 +344,62 @@ class CustomButton_main extends StatelessWidget {
           ),
         ),
       ),
+    );
+  }
+}
+
+class ListTile_main extends StatelessWidget {
+  final String text;
+  final String Menu_Icon;
+  final VoidCallback onTap;
+
+  const ListTile_main({
+    required this.text,
+    required this.Menu_Icon,
+    required this.onTap
+  });
+  @override
+  Widget build(BuildContext context){
+    return ListTile(
+        leading: SvgPicture.asset(Menu_Icon),
+        title: Text(
+          text,
+          style: TextStyle(
+            color: Colors.black,
+            fontFamily: 'Pretendard',
+            fontSize: 16.sp,
+            fontStyle: FontStyle.normal,
+            fontWeight: FontWeight.w400,
+            letterSpacing: -0.4.sp,
+          ),
+        ),
+        onTap: onTap,
+        trailing: Image.asset(
+          'asstes/Vector6.png',
+          scale: 4.sp,
+        )
+    );
+  }
+}
+
+class resign_Button extends StatelessWidget{
+  @override
+  Widget build(BuildContext context){
+    return TextButton(
+      onPressed: () {},
+      child: Text(
+        '회원 탈퇴하기 >',
+        style: TextStyle(
+          color: Color(0xff767676),
+          fontFamily: 'Pretendard',
+          fontSize: 12.sp,
+          fontStyle: FontStyle.normal,
+          fontWeight: FontWeight.w400,
+          letterSpacing: 0.3.sp,
+        ),
+      ),
+      style: TextButton.styleFrom(
+          padding: EdgeInsets.zero),
     );
   }
 }
